Rename: Change Project name
- Skeleton -> SocialMediaFeed
