+++ src/test/java/com/example/socialmediafeed/SocialMediaFeedApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.skeleton;
+package com.example.socialmediafeed;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SkeletonApplicationTests {
+class SocialMediaFeedApplicationTests {
 
     @Test
     void contextLoads() {
